Remove `RejectionReactor` and the related tests; use `RejectionSubscriber` instead, where applicable.
