+++ server/src/main/java/io/spine/server/rejection/RejectionSubscriber.java
@@ -19,30 +19,109 @@
  */
 package io.spine.server.rejection;
 
+import com.google.common.base.Supplier;
+import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Message;
+import io.spine.core.Command;
 import io.spine.core.CommandContext;
 import io.spine.core.RejectionClass;
+import io.spine.core.RejectionEnvelope;
 import io.spine.server.reflect.RejectionSubscriberMethod;
+import io.spine.server.tenant.CommandOperation;
+import io.spine.string.Stringifiers;
+import io.spine.type.MessageClass;
+import io.spine.util.Logging;
+import org.slf4j.Logger;
 
+import javax.annotation.Nullable;
 import java.util.Set;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+import static java.lang.String.format;
 
 /**
  * A base for objects subscribing to rejections from {@link RejectionBus}.
  *
- * @author Alex Tymchenko
  * @author Alexander Yevsyukov
+ * @author Alex Tymchenko
  * @see RejectionBus#register(io.spine.server.bus.MessageDispatcher)
  * @see io.spine.core.Subscribe
  */
-public class RejectionSubscriber extends AbstractRejectionDispatcher {
+public class RejectionSubscriber implements RejectionDispatcher<String> {
 
+    /**
+     * Cached set of the rejection classes this subscriber is subscribed to.
+     */
+    @Nullable
+    private Set<RejectionClass> rejectionClasses;
+
+    /** Lazily initialized logger. */
+    private final Supplier<Logger> loggerSupplier = Logging.supplyFor(getClass());
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param envelope the envelope with the message
+     * @return a one element set with the result of {@link #toString()} as the identity of
+     * the subscriber, or empty set if dispatching failed
+     */
     @Override
-    public void handle(Message rejectionMessage, Message commandMessage, CommandContext context) {
-        RejectionSubscriberMethod.invokeFor(this, rejectionMessage, commandMessage, context);
+    public Set<String> dispatch(final RejectionEnvelope envelope) {
+        final Command originCommand = envelope.getOuterObject()
+                                              .getContext()
+                                              .getCommand();
+        final CommandOperation op = new CommandOperation(originCommand) {
+
+            @Override
+            public void run() {
+                handle(envelope.getMessage(),
+                       envelope.getCommandMessage(),
+                       envelope.getCommandContext());
+            }
+        };
+        try {
+            op.execute();
+        } catch (RuntimeException e) {
+            onError(envelope, e);
+            return ImmutableSet.of();
+        }
+        return Identity.of(this);
+    }
+
+    /**
+     * Logs the error into the subscriber {@linkplain #log() log}.
+     *
+     * @param envelope  the message which caused the error
+     * @param exception the error
+     */
+    @Override
+    public void onError(RejectionEnvelope envelope, RuntimeException exception) {
+        checkNotNull(envelope);
+        checkNotNull(exception);
+        final MessageClass messageClass = envelope.getMessageClass();
+        final String messageId = Stringifiers.toString(envelope.getId());
+        final String errorMessage =
+                format("Error reacting on rejection (class: %s id: %s).", messageClass, messageId);
+        log().error(errorMessage, exception);
+    }
+
+    /**
+     * Obtains the instance of logger associated with the class of the subscriber.
+     */
+    protected Logger log() {
+        return loggerSupplier.get();
     }
 
     @Override
-    protected Set<RejectionClass> inspect() {
-        return RejectionSubscriberMethod.inspect(getClass());
+    @SuppressWarnings("ReturnOfCollectionOrArrayField") // as we return an immutable collection.
+    public Set<RejectionClass> getMessageClasses() {
+        if (rejectionClasses == null) {
+            rejectionClasses = RejectionSubscriberMethod.inspect(getClass());
+        }
+        return rejectionClasses;
+    }
+
+    public void handle(Message rejectionMessage, Message commandMessage, CommandContext context) {
+        RejectionSubscriberMethod.invokeFor(this, rejectionMessage, commandMessage, context);
     }
 }
